nit: use %+v consistently for formatting (#10)
Followup to https://github.com/peterldowns/testy/pull/9, this PR makes
it so that all `t.Error()` calls use `fmt.Sprintf()` with `%+v` to
format values. This prints them in their default representation, with
struct field names if the value is a struct, [as documented in the fmt
package](https://pkg.go.dev/fmt#hdr-Printing).

This is one of those weird little nitpicks I want to do to keep things
"consistent" it's not any more right or wrong than before.

## Testing

I added a new test that skips itself, but when un-skipped, prints all
the failure messages of the different checks. Running this test shows
that the messages fixed by #9 still print correctly, as do all other
messages:

```
--- FAIL: TestFailureMessages (0.00s)
    /Users/pd/code/testy/check/check_test.go:824: expected false
    /Users/pd/code/testy/check/check_test.go:825: expected true
    /Users/pd/code/testy/check/check_test.go:826: expected want != got
        want: [hello]
         got: [hello]
    /Users/pd/code/testy/check/check_test.go:827: expected want == got
        --- want
        +++ got
          map[string]int{
        + 	"goodbye": 2,
        - 	"hello":   1,
          }
        
    /Users/pd/code/testy/check/check_test.go:831: expected 1 > 4
    /Users/pd/code/testy/check/check_test.go:833: expected 8 < 6
    /Users/pd/code/testy/check/check_test.go:835: expected <nil> error, received oh no
    /Users/pd/code/testy/check/check_test.go:836: expected non-<nil> error, received <nil>
    /Users/pd/code/testy/check/check_test.go:837: expected slice to not contain element
        element: 4
          found: 4
    /Users/pd/code/testy/check/check_test.go:838: expected slice to contain element:
        element: hello
        
    /Users/pd/code/testy/check/check_test.go:841: expected non-<nil> value, received map[]
    /Users/pd/code/testy/check/check_test.go:843: expected <nil>, received map[hello:world]
FAIL
```